Use the correct math symbol for multiplication (#145)
It's good typographic practice to use the correct math symbols for
multiplication (see the links below).

https://practicaltypography.com/math-symbols.html

and

http://www.malinc.se/math/latex/basiccodeen.php

We also replace some hyphens with en-dashes where appropriate.

Note that both changes, use of the correct multiplication symbol as well
as the en-dash should be changed throughout the book for consistency. This
can be done if we decide to use these consistently.